+++ FlowDesigner/data-flow/src/add_operators.cc
@@ -34,7 +34,7 @@ ObjectRef addVectorFunction(ObjectRef op1, ObjectRef op2) {
   //creating new vector
   RCPtr<Z> resultValue(Z::alloc(op1Value->size()));
 
-  for (int i = 0; i < resultValue->size(); i++) {    
+  for (size_t i = 0; i < resultValue->size(); i++) {    
     (*resultValue)[i] = static_cast<typename Z::basicType> ((*op1Value)[i]) + static_cast<typename Z::basicType> ((*op2Value)[i]);
   }
 
@@ -56,8 +56,8 @@ ObjectRef addMatrixFunction(ObjectRef op1, ObjectRef op2) {
   
   RCPtr<Z> resultValue(new Z(op1Value->nrows(), op1Value->ncols()));
   
-  for (int i = 0; i < resultValue->nrows(); i++) {    
-    for (int j = 0; j < resultValue->ncols(); j++) {      
+  for (size_t i = 0; i < resultValue->nrows(); i++) {    
+    for (size_t j = 0; j < resultValue->ncols(); j++) {      
       (*resultValue)(i,j) = static_cast<typename Z::basicType> ((*op1Value)(i,j)) + static_cast<typename Z::basicType> ((*op2Value)(i,j));    
      }
   }
@@ -78,7 +78,7 @@ ObjectRef addVectorObjectRef(ObjectRef op1, ObjectRef op2) {
   //creating new vector
   RCPtr<Vector<ObjectRef> > resultValue(new Vector<ObjectRef>(op1Value->size()));
 
-  for (int i = 0; i < resultValue->size(); i++) {    
+  for (size_t i = 0; i < resultValue->size(); i++) {    
     (*resultValue)[i] = (*op1Value)[i] + (*op2Value)[i];
   }
 
@@ -98,8 +98,8 @@ ObjectRef addMatrixObjectRef(ObjectRef op1, ObjectRef op2) {
   //creating new Matrix
   RCPtr<Matrix<ObjectRef> > resultValue(new Matrix<ObjectRef>(op1Value->nrows(), op1Value->ncols()));
 
-  for (int i = 0; i < resultValue->nrows(); i++) {    
-    for (int j = 0; j < resultValue->ncols(); j++) {      
+  for (size_t i = 0; i < resultValue->nrows(); i++) {    
+    for (size_t j = 0; j < resultValue->ncols(); j++) {      
       (*resultValue)(i,j) = (*op1Value)(i,j) + (*op2Value)(i,j);    
     }
   }
@@ -117,7 +117,7 @@ ObjectRef addVectorScalarFunction(ObjectRef op1, ObjectRef op2) {
   //creating new vector
   RCPtr<Z> resultValue(Z::alloc(op1Value->size()));
 
-  for (int i = 0; i < resultValue->size(); i++) {    
+  for (size_t i = 0; i < resultValue->size(); i++) {    
     (*resultValue)[i] = static_cast<typename Z::basicType> ((*op1Value)[i]) + static_cast<typename Z::basicType> (op2Value->val());
   }
 
@@ -137,8 +137,8 @@ ObjectRef addMatrixScalarFunction(ObjectRef op1, ObjectRef op2) {
   
   RCPtr<Z> resultValue(new Z(op1Value->nrows(), op1Value->ncols()));
   
-  for (int i = 0; i < resultValue->nrows(); i++) {    
-    for (int j = 0; j < resultValue->ncols(); j++) {      
+  for (size_t i = 0; i < resultValue->nrows(); i++) {    
+    for (size_t j = 0; j < resultValue->ncols(); j++) {      
       (*resultValue)(i,j) = static_cast<typename Z::basicType> ((*op1Value)(i,j)) + static_cast<typename Z::basicType> (op2Value->val());    
      }
   }
